+++ src/actions.rs
@@ -204,10 +204,7 @@ pub fn edit(
     notes: Option<&str>,
     folder_uuid: Option<&str>,
     history: &[crate::db::HistoryEntry],
-    key: Option<&str>,
-    reprompt: crate::api::CipherRepromptType,
-    favorite: bool,
-    archived_date: Option<&str>,
+    meta: &crate::db::CipherWriteMeta<'_>,
 ) -> Result<(Option<String>, ())> {
     with_exchange_refresh_token(access_token, refresh_token, |access_token| {
         edit_once(
@@ -220,10 +217,7 @@ pub fn edit(
             notes,
             folder_uuid,
             history,
-            key,
-            reprompt,
-            favorite,
-            archived_date,
+            meta,
         )
     })
 }
@@ -238,10 +232,7 @@ fn edit_once(
     notes: Option<&str>,
     folder_uuid: Option<&str>,
     history: &[crate::db::HistoryEntry],
-    key: Option<&str>,
-    reprompt: crate::api::CipherRepromptType,
-    favorite: bool,
-    archived_date: Option<&str>,
+    meta: &crate::db::CipherWriteMeta<'_>,
 ) -> Result<()> {
     let (client, _) = api_client()?;
     client.edit(
@@ -254,10 +245,7 @@ fn edit_once(
         notes,
         folder_uuid,
         history,
-        key,
-        reprompt,
-        favorite,
-        archived_date,
+        meta,
     )?;
     Ok(())
 }